requirements-freeze.txt
apsw==3.16.2.post1
asn1crypto==0.24.0
av==6.1.2
beautifulsoup4==4.6.0
chardet==3.0.4
CherryPy==8.9.1
Click==7.0
cmake==3.13.3
cryptography==2.1.4
cssselect==1.0.3
cssutils==1.0.2
defusedxml==0.5.0
dlib==19.16.0
dnspython==1.15.0
enum34==1.1.6
face-recognition==1.2.3
face-recognition-models==0.3.0
feedparser==5.2.1
gyp==0.1
html5-parser==0.4.4
html5lib==0.999999999
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
lxml==4.2.1
Markdown==2.6.9
mechanize==0.2.5
msgpack==0.5.6
netifaces==0.10.4
numpy==1.16.1
olefile==0.45.1
opencv-python==4.0.0.21
Pillow==5.4.1
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
pyliblzma==0.5.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pyxdg==0.25
PyYAML==3.12
regex==2017.12.12
repoze.lru==0.7
Routes==2.4.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
SOAPpy==0.12.22
uTidylib==0.3
virtualenv==16.1.0
webencodings==0.5
WebOb==1.7.3
wstools==0.4.3
